Make gcp-routes be path-activated rather than polling files
For historical reasons, gcp-routes was a persistent service that
polled for changes to /run/cloud-routes files. When azure-routes was
written, it was improved by making it use systemd+inotify instead but
that improvement was never backported to gcp-routes. Fix that.

(Also remove some unnecessary conditions from the gcp-routes unit
file: we only install this file on GCP, so we don't need to
conditionalize the systemd unit on the platform as well.)